// src/lib/types/types.ts
type CustomerProfileProps = {
  id: string;
  name: string;
  email: string;
  gender: "male" | "female";
  phone: string;
  image?: string;
  createdAt: string;
  updatedAt: string;
  lastLogin: string;
};